feat(editor): reorganize page components, use router search service
