+++ apps/metadata-editor/src/app/records/all-records/all-records-list.component.html
@@ -1,36 +1,2 @@
-<main class="grow bg-gray-100 p-12 flex flex-col justify-between gap-12">
-  <div class="flex justify-center gap-8 items-baseline">
-    <h1 class="text-[56px] capitalize font-title grow" translate="">catalog</h1>
-    <button
-      role="button"
-      class="btn-default flex gap-3 items-center"
-      (click)="createRecord()"
-    >
-      <mat-icon class="material-icons-outlined">post_add</mat-icon>
-      <span translate="">new record</span>
-    </button>
-  </div>
-  <div
-    class="shadow rounded bg-white grow"
-    *ngIf="searchFacade.results$ | async as results"
-  >
-    <gn-ui-record-table
-      [records]="results"
-      [totalHits]="searchFacade.resultsHits$ | async"
-      (recordSelect)="editRecord($event)"
-    ></gn-ui-record-table>
-    <div
-      class="px-5 py-5 flex justify-center gap-8 items-baseline"
-      *ngIf="searchFacade.resultsHits$ | async as hits"
-    >
-      <div class="grow">
-        <gn-ui-pagination-buttons
-          hideButton="true"
-          [currentPage]="searchFacade.currentPage$ | async"
-          [totalPages]="searchFacade.totalPages$ | async"
-          (newCurrentPageEvent)="paginate($event)"
-        ></gn-ui-pagination-buttons>
-      </div>
-    </div>
-  </div>
-</main>
+<md-editor-records-list [title]="'catalog' | translate">
+</md-editor-records-list>
