+++ apps/metadata-editor/src/app/records/my-draft/my-draft.component.html
@@ -1,1 +1,2 @@
-<h1 class="p-5">My Draft page works!</h1>
+<md-editor-records-list [title]="'my draft' | translate">
+</md-editor-records-list>
