+++ apps/metadata-editor/src/app/records/my-library/my-library.component.html
@@ -1,1 +1,2 @@
-<h1 class="p-5">My Library page works!</h1>
+<md-editor-records-list [title]="'results.for.myOrg' | translate">
+</md-editor-records-list>
