+++ apps/metadata-editor/src/app/records/my-org-records/my-org-records.component.html
@@ -1,1 +1,2 @@
-<h1 class="p-5">My Organisation Records page works!</h1>
+<md-editor-records-list [title]="'my organization records' | translate">
+</md-editor-records-list>
