+++ apps/metadata-editor/src/app/records/my-org-records/my-org-records.component.spec.ts
@@ -1,15 +1,46 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 
 import { MyOrgRecordsComponent } from './my-org-records.component'
+import { SearchFacade } from '@geonetwork-ui/feature/search'
+import { Component, importProvidersFrom } from '@angular/core'
+import { TranslateModule } from '@ngx-translate/core'
+import { RecordsListComponent } from '../records-list.component'
+
+@Component({
+  // eslint-disable-next-line
+  selector: 'md-editor-records-list',
+  template: '',
+  standalone: true,
+})
+export class MockRecordsListComponent {}
+
+class SearchFacadeMock {
+  setFilters = jest.fn()
+}
 
 describe('MyOrgRecordsComponent', () => {
   let component: MyOrgRecordsComponent
   let fixture: ComponentFixture<MyOrgRecordsComponent>
+  let searchFacade: SearchFacade
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [MyOrgRecordsComponent],
+      providers: [
+        importProvidersFrom(TranslateModule.forRoot()),
+        {
+          provide: SearchFacade,
+          useClass: SearchFacadeMock,
+        },
+      ],
+    }).overrideComponent(MyOrgRecordsComponent, {
+      remove: {
+        imports: [RecordsListComponent],
+      },
+      add: {
+        imports: [MockRecordsListComponent],
+      },
     })
+    searchFacade = TestBed.inject(SearchFacade)
     fixture = TestBed.createComponent(MyOrgRecordsComponent)
     component = fixture.componentInstance
     fixture.detectChanges()
@@ -17,5 +48,11 @@ describe('MyOrgRecordsComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy()
+  })
+
+  describe('filters', () => {
+    it('clears filters on init', () => {
+      expect(searchFacade.setFilters).toHaveBeenCalledWith({})
+    })
   })
 })
