+++ apps/metadata-editor/src/app/records/my-records/my-records.component.html
@@ -1,1 +1,2 @@
-<h1 class="p-5">My Records page works!</h1>
+<md-editor-records-list [title]="'my records' | translate">
+</md-editor-records-list>
