+++ apps/metadata-editor/src/app/records/records-list.component.html
@@ -1,0 +1,27 @@
+<main class="grow bg-gray-100 p-12 flex flex-col justify-between gap-12">
+  <div class="flex justify-center gap-8 items-baseline">
+    <h1 class="text-[56px] capitalize font-title grow">{{ title }}</h1>
+  </div>
+  <div
+    class="shadow rounded bg-white grow"
+    *ngIf="searchFacade.results$ | async as results"
+  >
+    <gn-ui-record-table
+      [records]="results"
+      [totalHits]="searchFacade.resultsHits$ | async"
+      (recordSelect)="editRecord($event)"
+    ></gn-ui-record-table>
+    <div
+      class="px-5 py-5 flex justify-center gap-8 items-baseline"
+      *ngIf="searchFacade.resultsHits$ | async as hits"
+    >
+      <div class="grow">
+        <gn-ui-pagination-buttons
+          [currentPage]="searchFacade.currentPage$ | async"
+          [totalPages]="searchFacade.totalPages$ | async"
+          (newCurrentPageEvent)="paginate($event)"
+        ></gn-ui-pagination-buttons>
+      </div>
+    </div>
+  </div>
+</main>
