+++ apps/metadata-editor/src/app/records/records-list.component.ts
@@ -1,0 +1,51 @@
+import { CommonModule } from '@angular/common'
+import { Component, Input } from '@angular/core'
+import { MatIconModule } from '@angular/material/icon'
+import { Router } from '@angular/router'
+import { CatalogRecord } from '@geonetwork-ui/common/domain/record'
+import { SearchFacade, SearchService } from '@geonetwork-ui/feature/search'
+import { UiSearchModule } from '@geonetwork-ui/ui/search'
+import { UiElementsModule } from '@geonetwork-ui/ui/elements'
+
+const includes = [
+  'uuid',
+  'resourceTitleObject',
+  'createDate',
+  'changeDate',
+  'userinfo',
+  'cl_status',
+  'isPublishedToAll',
+]
+
+@Component({
+  selector: 'md-editor-records-list',
+  templateUrl: './records-list.component.html',
+  styleUrls: ['./records-list.component.css'],
+  standalone: true,
+  imports: [CommonModule, MatIconModule, UiSearchModule, UiElementsModule],
+})
+export class RecordsListComponent {
+  @Input() title: string
+
+  constructor(
+    private router: Router,
+    public searchFacade: SearchFacade,
+    public searchService: SearchService
+  ) {
+    this.searchFacade
+      .setPagination(0, 15)
+      .setConfigRequestFields(includes)
+      .setSortBy(['desc', 'changeDate'])
+  }
+
+  paginate(page: number) {
+    this.searchService.setPage(page)
+  }
+  createRecord() {
+    this.router.navigate(['/create'])
+  }
+
+  editRecord(record: CatalogRecord) {
+    this.router.navigate(['/edit', record.uniqueIdentifier])
+  }
+}
