+++ apps/metadata-editor/src/app/records/search-records/search-records-list.component.html
@@ -1,58 +1,2 @@
-<main class="grow bg-gray-100 p-12 flex flex-col justify-between gap-12">
-  <div class="flex justify-center gap-8 items-baseline">
-    <h1 class="text-[56px] normal-case font-title grow">
-      <ng-container
-        *ngrxLet="searchFacade.searchFilters$ | async as searchFilters"
-      >
-        <ng-container *ngIf="searchFilters['any']"
-          >{{ 'results.for.any' | translate }} "{{
-            searchFilters['any']
-          }}"</ng-container
-        >
-        <ng-container *ngIf="!searchFilters['any']">{{
-          'catalog' | translate
-        }}</ng-container>
-      </ng-container>
-    </h1>
-    <button
-      role="button"
-      class="btn-default flex gap-3 items-center"
-      (click)="createRecord()"
-    >
-      <mat-icon class="material-icons-outlined">post_add</mat-icon>
-      <span translate="">new record</span>
-    </button>
-  </div>
-  <div
-    class="shadow rounded bg-white grow"
-    *ngIf="searchFacade.results$ | async as results"
-  >
-    <gn-ui-record-table
-      [records]="results"
-      (recordSelect)="editRecord($event)"
-    ></gn-ui-record-table>
-    <div
-      class="px-5 py-5 flex justify-center gap-8 items-baseline"
-      *ngIf="searchFacade.resultsHits$ | async as hits"
-    >
-      <div
-        translate
-        class="text-gray-500"
-        [translateParams]="{
-          displayed: results.length,
-          hits: hits.value
-        }"
-      >
-        results.records.hits.displayedOn
-      </div>
-      <div class="grow">
-        <gn-ui-pagination
-          hideButton="true"
-          [currentPage]="searchFacade.currentPage$ | async"
-          [nPages]="searchFacade.totalPages$ | async"
-          (newCurrentPageEvent)="paginate($event)"
-        ></gn-ui-pagination>
-      </div>
-    </div>
-  </div>
-</main>
+<md-editor-records-list [title]="'results.for.any' | translate">
+</md-editor-records-list>
