+++ apps/metadata-editor/src/app/records/search-records/search-records-list.component.ts
@@ -1,48 +1,16 @@
 import { CommonModule } from '@angular/common'
 import { Component } from '@angular/core'
-import { MatIconModule } from '@angular/material/icon'
-import { Router } from '@angular/router'
-import { CatalogRecord } from '@geonetwork-ui/common/domain/record'
-import { SearchFacade, SearchService } from '@geonetwork-ui/feature/search'
-import { UiSearchModule } from '@geonetwork-ui/ui/search'
-
-const includes = [
-  'uuid',
-  'resourceTitleObject',
-  'createDate',
-  'changeDate',
-  'userinfo',
-  'cl_status',
-  'isPublishedToAll',
-]
+import { SearchFacade } from '@geonetwork-ui/feature/search'
+import { RecordsListComponent } from '../records-list.component'
+import { TranslateModule } from '@ngx-translate/core'
 
 @Component({
   selector: 'md-editor-search-records-list',
   templateUrl: './search-records-list.component.html',
   styleUrls: ['./search-records-list.component.css'],
   standalone: true,
-  imports: [CommonModule, MatIconModule, UiSearchModule],
+  imports: [CommonModule, TranslateModule, RecordsListComponent],
 })
 export class SearchRecordsComponent {
-  constructor(
-    private router: Router,
-    public searchFacade: SearchFacade,
-    public searchService: SearchService
-  ) {
-    this.searchFacade
-      .setConfigRequestFields(includes)
-      .setSortBy(['desc', 'changeDate'])
-      .setPagination(0, 15)
-  }
-
-  paginate(page: number) {
-    this.searchService.setPage(page)
-  }
-  createRecord() {
-    this.router.navigate(['/create'])
-  }
-
-  editRecord(record: CatalogRecord) {
-    this.router.navigate(['/edit', record.uniqueIdentifier])
-  }
+  constructor(public searchFacade: SearchFacade) {}
 }
