+++ apps/metadata-editor/src/app/router.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core'
 import { Router } from '@angular/router'
-import { ROUTER_ROUTE_SEARCH } from '@geonetwork-ui/feature/router'
 import { appRoutes } from './app.routes'
 
 @Injectable({
@@ -14,6 +13,6 @@ export class EditorRouterService {
   }
 
   getSearchRoute(): string {
-    return ROUTER_ROUTE_SEARCH
+    return 'records/search'
   }
 }
